Разделил SqlitePhotoRepository на SqlitePhotoCommandRepository и SqlitePhotoQueryRepository
