+++ src/PhotoCatalog.Infrastructure/Repositories/SqlitePhotoQueryRepository.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -7,7 +7,6 @@ using Dapper;
 using Microsoft.Data.Sqlite;
 
 using PhotoCatalog.Domain.Entities;
-using PhotoCatalog.Domain.Interfaces.Repositories;
 using PhotoCatalog.Domain.Primitives;
 using PhotoCatalog.Infrastructure.Errors;
 using PhotoCatalog.Infrastructure.UnitOfWork;
@@ -17,9 +16,9 @@ using Serilog;
 namespace PhotoCatalog.Infrastructure.Repositories;
 
 /// <summary>
-///     Реализация репозитория для работы с фотографиями в SQLite с использованием Dapper.
+///     Реализация репозитория для работы с получением фотографий в SQLite с использованием Dapper.
 /// </summary>
-public class SqlitePhotoRepository : IPhotoCommandRepository
+public class SqlitePhotoQueryRepository
 {
     private readonly SqliteUnitOfWork _unitOfWork;
     private readonly ILogger _logger;
@@ -29,7 +28,7 @@ public class SqlitePhotoRepository : IPhotoCommandRepository
     /// </summary>
     /// <param name="unitOfWork">Экземпляр Unit of Work для доступа к соединению и транзакции.</param>
     /// <param name="logger">Логгер для записи ошибок.</param>
-    public SqlitePhotoRepository(SqliteUnitOfWork unitOfWork, ILogger logger)
+    public SqlitePhotoQueryRepository(SqliteUnitOfWork unitOfWork, ILogger logger)
     {
         _unitOfWork = unitOfWork;
         _logger = logger;
@@ -124,176 +123,6 @@ public class SqlitePhotoRepository : IPhotoCommandRepository
         {
             _logger.Error(ex, "Неожиданная ошибка в методе GetByPath для пути {RealPath}", realPath);
             return Result<Photo>.Failure(InfrastructureErrors.Database.ConnectionFailed);
-        }
-    }
-
-    /// <summary>
-    ///     Добавляет новую фотографию в репозиторий.
-    /// </summary>
-    /// <param name="photo">Объект фотографии для добавления.</param>
-    /// <returns>Результат операции.</returns>
-    public ResultVoid Add(Photo photo)
-    {
-        try
-        {
-            var connection = _unitOfWork.Connection;
-            if (connection == null)
-            {
-                return ResultVoid.Failure(InfrastructureErrors.Database.ConnectionFailed);
-            }
-
-            var dimensionsValue = $"{photo.Dimensions.Width}x{photo.Dimensions.Height}";
-
-            connection.Execute(
-                "INSERT INTO Photos (Id, RealPath, FileHash, Dimensions, AddedAt) VALUES (@Id, @RealPath, @FileHash, @Dimensions, @AddedAt)",
-                new
-                {
-                    photo.Id,
-                    photo.RealPath,
-                    photo.FileHash,
-                    Dimensions = dimensionsValue,
-                    photo.AddedAt
-                },
-                _unitOfWork.Transaction);
-
-            var tagValues = photo.TagIds.Select(tagId => new { PhotoId = photo.Id, TagId = tagId });
-
-            connection.Execute(
-                "INSERT INTO PhotoTags (PhotoId, TagId) VALUES (@PhotoId, @TagId)",
-                tagValues,
-                _unitOfWork.Transaction);
-
-
-            return ResultVoid.Success();
-        }
-        catch (SqliteException ex) when (ex.SqliteErrorCode == 19)
-        {
-            _logger.Error(ex, "Нарушение уникальности RealPath для фотографии {RealPath}", photo.RealPath);
-            return ResultVoid.Failure(InfrastructureErrors.Database.ConstraintViolation);
-        }
-        catch (SqliteException ex)
-        {
-            _logger.Error(ex, "Ошибка SQLite в методе Add для фотографии {Id}", photo.Id);
-            return ResultVoid.Failure(InfrastructureErrors.Database.ConnectionFailed);
-        }
-        catch (Exception ex)
-        {
-            _logger.Error(ex, "Неожиданная ошибка в методе Add для фотографии {Id}", photo.Id);
-            return ResultVoid.Failure(InfrastructureErrors.Database.ConnectionFailed);
-        }
-    }
-
-    /// <summary>
-    ///     Обновляет существующую фотографию.
-    /// </summary>
-    /// <param name="photo">Объект фотографии с обновленными данными.</param>
-    /// <returns>Результат операции.</returns>
-    public ResultVoid Update(Photo photo)
-    {
-        if (photo == null)
-        {
-            return ResultVoid.Failure(DomainErrors.Photo.NullPhoto);
-        }
-
-        try
-        {
-            var connection = _unitOfWork.Connection;
-            if (connection == null)
-            {
-                return ResultVoid.Failure(InfrastructureErrors.Database.ConnectionFailed);
-            }
-
-            var dimensionsValue = $"{photo.Dimensions.Width}x{photo.Dimensions.Height}";
-
-            var rowsAffected = connection.Execute(
-                "UPDATE Photos SET RealPath = @RealPath, FileHash = @FileHash, Dimensions = @Dimensions, AddedAt = @AddedAt WHERE Id = @Id",
-                new
-                {
-                    photo.Id,
-                    photo.RealPath,
-                    photo.FileHash,
-                    Dimensions = dimensionsValue,
-                    photo.AddedAt
-                },
-                _unitOfWork.Transaction);
-
-            if (rowsAffected == 0)
-            {
-                return ResultVoid.Failure(DomainErrors.Photo.NotFound);
-            }
-
-            connection.Execute(
-                "DELETE FROM PhotoTags WHERE PhotoId = @PhotoId",
-                new { PhotoId = photo.Id },
-                _unitOfWork.Transaction);
-
-            var tagValues = photo.TagIds.Select(tagId => new { PhotoId = photo.Id, TagId = tagId });
-
-            connection.Execute(
-                "INSERT INTO PhotoTags (PhotoId, TagId) VALUES (@PhotoId, @TagId)",
-                tagValues,
-                _unitOfWork.Transaction);
-
-            return ResultVoid.Success();
-        }
-        catch (SqliteException ex) when (ex.SqliteErrorCode == 19)
-        {
-            _logger.Error(ex, "Нарушение уникальности RealPath для фотографии {Id}", photo.Id);
-            return ResultVoid.Failure(InfrastructureErrors.Database.ConstraintViolation);
-        }
-        catch (SqliteException ex)
-        {
-            _logger.Error(ex, "Ошибка SQLite в методе Update для фотографии {Id}", photo.Id);
-            return ResultVoid.Failure(InfrastructureErrors.Database.ConnectionFailed);
-        }
-        catch (Exception ex)
-        {
-            _logger.Error(ex, "Неожиданная ошибка в методе Update для фотографии {Id}", photo.Id);
-            return ResultVoid.Failure(InfrastructureErrors.Database.ConnectionFailed);
-        }
-    }
-
-    /// <summary>
-    ///     Удаляет фотографию по идентификатору.
-    /// </summary>
-    /// <param name="id">Идентификатор фотографии для удаления.</param>
-    /// <returns>Результат операции.</returns>
-    public ResultVoid Delete(int id)
-    {
-        try
-        {
-            var connection = _unitOfWork.Connection;
-            if (connection == null)
-            {
-                return ResultVoid.Failure(InfrastructureErrors.Database.ConnectionFailed);
-            }
-
-            connection.Execute(
-                "DELETE FROM PhotoTags WHERE PhotoId = @Id",
-                new { Id = id },
-                _unitOfWork.Transaction);
-
-            var rowsAffected = connection.Execute(
-                "DELETE FROM Photos WHERE Id = @Id",
-                new { Id = id },
-                _unitOfWork.Transaction);
-
-            if (rowsAffected == 0)
-            {
-                return ResultVoid.Failure(DomainErrors.Photo.NotFound);
-            }
-
-            return ResultVoid.Success();
-        }
-        catch (SqliteException ex)
-        {
-            _logger.Error(ex, "Ошибка SQLite в методе Delete для фотографии {Id}", id);
-            return ResultVoid.Failure(InfrastructureErrors.Database.ConnectionFailed);
-        }
-        catch (Exception ex)
-        {
-            _logger.Error(ex, "Неожиданная ошибка в методе Delete для фотографии {Id}", id);
-            return ResultVoid.Failure(InfrastructureErrors.Database.ConnectionFailed);
         }
     }
